Fix httpx.ResponseNotRead if streaming unready pod logs
If calling Client.log() on a Pod that exists but the container has not
started yet, the Kubernetes API will return 400, considering that a
client error.

Lightkube will try to convert that httpx.HTTPStatusError into an
ApiError. While doing so, it will read the response body for the Status
object that the Kubernetes API returns on errors.

However, if the `follow` parameter is True, this function will set
`stream` to True, which will instruct httpx to delay reading of the
response. Then, when lightkube attempts to parse the response body,
httpx errors because it has not yet been read.

To fix this, simply read the body before creating the error. Now an
ApiError will be raised with the status set to 400 as expected.